zp working in cuda in 1d

+++ CUDA_programming/zeropadding_rough_work/zeropadding_C_1.1/zp_c_ctypes_1d_v2.c
@@ -10,13 +10,14 @@ void zeroPad(
     double* in_array,
     double* out_array,
     long* edges,
-    long n_blocks,
-    long largest_block
+    int n_blocks,
+    int largest_block
 ){
     for (int block = 0; block < n_blocks; block++){
-        long start = edges[block];
-        long stop = edges[block + 1];
-        long block_size = stop - start;
+        // printf("Test 1");
+        int start = edges[block];
+        int stop = edges[block + 1];
+        int block_size = stop - start;
         for (int i = 0; i < block_size; i++){
             out_array[block*largest_block + i] = in_array[start + i];
         }
